[eg] add: mount_up() func

+++ bot_utils/picture_in_picture/picture_input/picture_input_abc.py
@@ -4,12 +4,13 @@ from bot_utils.utils import DebugAbstractClass
 
 
 class PictureInputAbstractClass(DebugAbstractClass):
-    def __init__(self, *args, **kwargs):
+    def __init__(self, cleanup_after=True, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.img_path = None
+        self.cleanup_after = cleanup_after
 
     def get_image(self, img_path):
-        self.img_path = img_path
+        self.img_path = os.path.abspath(img_path)
 
     def update_region(self, region):
         """if the picture input implementated class does something to offset pixels
@@ -22,5 +23,5 @@ class PictureInputAbstractClass(DebugAbstractClass):
         return region
 
     def clean(self):
-        if self.img_path:
+        if self.img_path and self.cleanup_after:
             os.remove(self.img_path)
